Added creation and deletion and edit tests for asset. Refactored other tests like chatboard tests. Removed unused imports and added some null checks for various functions.

+++ project/src/main/java/cs2815/project/service/Implementations/ChatServiceImpl.java
@@ -46,6 +46,7 @@ public class ChatServiceImpl implements ChatService {
      */
     @Override
     public void deleteMessage(ChatBoard message) {
+        if(message == null){return;}
         chatRepo.delete(message);
     }
